requirements.txt
modal
PyYAML
httpx